+++ apps/client/src/app/core/database/storage/list/firestore-list-storage.ts
@@ -12,6 +12,7 @@ import { FirestoreRelationalStorage } from '../firestore/firestore-relational-st
 import { ListTag } from '../../../../modules/list/model/list-tag.enum';
 import { Class } from '@kaiu/serializer';
 import * as firebase from 'firebase/app';
+import { PermissionLevel } from '../../permissions/permission-level.enum';
 
 @Injectable({
   providedIn: 'root'
@@ -121,7 +122,7 @@ export class FirestoreListStorage extends FirestoreRelationalStorage<List> imple
   }
 
   public getShared(userId: string): Observable<List[]> {
-    return this.firestore.collection(this.getBaseUri(), ref => ref.where(`registry.${userId}`, '>=', 20))
+    return this.firestore.collection(this.getBaseUri(), ref => ref.where(`registry.${userId}`, '>=', PermissionLevel.READ))
       .snapshotChanges()
       .pipe(
         switchMap((snaps: DocumentChangeAction<List>[]) => {