+++ apps/client/src/app/modules/list/+state/lists.facade.ts
@@ -80,7 +80,7 @@ export class ListsFacade {
           map(([compacts, userId]) => {
             return compacts.filter(c => {
               return !c.notFound
-                && c.getPermissionLevel(userId) >= PermissionLevel.PARTICIPATE
+                && c.getPermissionLevel(userId) >= PermissionLevel.READ
                 && c.hasExplicitPermissions(userId)
                 && c.authorId !== userId;
             });
@@ -99,7 +99,7 @@ export class ListsFacade {
           return this.sortLists(
             compacts.filter(c => {
               return !c.notFound
-                && Math.max(c.getPermissionLevel(userId), c.getPermissionLevel(fcId)) >= PermissionLevel.PARTICIPATE
+                && Math.max(c.getPermissionLevel(userId), c.getPermissionLevel(fcId)) >= PermissionLevel.READ
                 && (c.hasExplicitPermissions(userId) || c.hasExplicitPermissions(fcId))
                 && c.authorId !== userId;
             })
@@ -163,7 +163,7 @@ export class ListsFacade {
       }
       let teamPermissionLevel = 0;
       if (team !== undefined && list.teamId === team.$key) {
-        teamPermissionLevel = list.getPermissionLevel(`team:${list.teamId}`) || 20;
+        teamPermissionLevel = Math.max(list.getPermissionLevel(`team:${list.teamId}`), 20);
       }
       return Math.max(list.getPermissionLevel(userId), list.getPermissionLevel(fcId), teamPermissionLevel);
     }),