+++ apps/client/src/environments/game-env.ts
@@ -1,6 +1,6 @@
 export const gameEnv = {
   maxLevel: 80,
   globalGameVersion: 5.25,
-  koreanGameVersion: 5.15,
-  chineseGameVersion: 5.18
+  koreanGameVersion: 5.18,
+  chineseGameVersion: 5.2
 };