+++ desktop/machina.js
@@ -36,7 +36,9 @@ module.exports.start = function(win, config, verbose, winpcap, pid) {
     log.info('elevated', elevated);
     if (elevated) {
       exec('netsh advfirewall firewall delete rule name="FFXIVTeamcraft"', () => {
-        exec(`netsh advfirewall firewall add rule name="FFXIVTeamcraft" dir=in action=allow program="${machinaExePath}" enable=yes`);
+        exec('netsh advfirewall firewall delete rule name="ffxiv teamcraft.exe"', () => {
+          exec(`netsh advfirewall firewall add rule name="FFXIVTeamcraft" dir=in action=allow program="${machinaExePath}" enable=yes`);
+        });
       });
 
       const region = config.get('region', null);